This edit solves the inter-file connectivity issues through corrected linkages
